Allow func type param names in the field name subsection
It is very natural to extend the field name subsection to include
parameters for function types. The binary encoding for function types
would be identical, and func types are already included under "composite
types" in the spec anyway.

In this patch I have updated to keep the name "field index" and "field
name subsection" because "field" is the more natural name to use in most
places in the spec. Renaming everything to "composite index" felt
dramatically worse to me.